Correction of addon on Cellphone
In GameLoop.cpp
Change the parameter of resetGrid, gameOver and whoWins to pointer.
Correction of condition in whoWins.
Change condition of reset in OnLoop.
In GameLoop
Correction of the grid type.
Correction of parameter of ResetGrid, whoWins and GameOver
In GL_Event
Comment mouse click
In GL_Init
Render loading screen
Change loading path